Added python script to make manifest file(s) and a metadata file for qiime
This script does not use the studyMetadata file, as I figured that all the fields in there were the same across samples/libraries in a qiime analysis, and therefore not as informative. If a user already has filled out these templates, then they have all of that study/assay information already formatted for publishing or meta-analysis. But the script can be changed to take that file in as well.